+++ packages/tswebextension/src/lib/mv2/background/services/csp-service.ts
@@ -1,8 +1,15 @@
 import { getDomain } from 'tldts';
 import { nanoid } from 'nanoid';
+import { NetworkRuleOption } from '@adguard/tsurlfilter';
 
-import { ContentType } from '../../../common/request-type';
-import { defaultFilteringLog, FilteringEventType } from '../../../common/filtering-log';
+import {
+    ContentType,
+} from '../../../common/request-type';
+import {
+    defaultFilteringLog,
+    FilteringEventType,
+    type FilteringLogInterface,
+} from '../../../common/filtering-log';
 import {
     type RequestContext,
     requestContextStorage,
@@ -16,13 +23,27 @@ const CSP_HEADER_NAME = 'Content-Security-Policy';
  */
 export class CspService {
     /**
+     * Filtering log.
+     */
+    private filteringLog: FilteringLogInterface;
+
+    /**
+     * Constructor.
+     *
+     * @param filteringLog Filtering log.
+     */
+    constructor(filteringLog: FilteringLogInterface) {
+        this.filteringLog = filteringLog;
+    }
+
+    /**
      * Applies CSP rules to response headers and returns modified headers.
      * It is applied when webRequest.onHeadersReceived event is fired.
      *
      * @param context Request context.
      * @returns True if headers were modified.
      */
-    public static onHeadersReceived(context: RequestContext): boolean {
+    public onHeadersReceived(context: RequestContext): boolean {
         const {
             matchingResult,
             responseHeaders,
@@ -42,6 +63,12 @@ export class CspService {
 
         for (let i = 0; i < cspRules.length; i += 1) {
             const rule = cspRules[i];
+            if (rule.isOptionEnabled(NetworkRuleOption.Header)) {
+                const responseHeaderMatch = rule.matchResponseHeaders(responseHeaders);
+                if (!responseHeaderMatch || rule.isAllowlist()) {
+                    continue;
+                }
+            }
             // Don't forget: getCspRules returns all $csp rules, we must directly check that the rule is blocking.
             if (RequestBlockingApi.isRequestBlockedByRule(rule)) {
                 const cspHeaderValue = rule.getAdvancedModifierValue();
@@ -52,7 +79,7 @@ export class CspService {
                         value: cspHeaderValue,
                     });
 
-                    defaultFilteringLog.publishEvent({
+                    this.filteringLog.publishEvent({
                         type: FilteringEventType.ApplyCspRule,
                         data: {
                             tabId,
@@ -80,3 +107,5 @@ export class CspService {
         return false;
     }
 }
+
+export const cspService = new CspService(defaultFilteringLog);