+++ packages/tswebextension/test/lib/mv2/background/services/remove-headers-service.test.ts
@@ -1,48 +1,53 @@
 import { MatchingResult, NetworkRule, RequestType } from '@adguard/tsurlfilter';
-import { HeadersService } from '@lib/mv2/background/services/headers-service';
+import { RemoveHeadersService } from '@lib/mv2/background/services/remove-headers-service';
 import { RequestContext } from '@lib/mv2/background/request';
 import { FilteringEventType, ContentType } from '@lib/common';
 import { MockFilteringLog } from '../../../common/mocks/mock-filtering-log';
 
 describe('Headers service', () => {
     const mockFilteringLog = new MockFilteringLog();
-    const headersService = new HeadersService(mockFilteringLog);
+    const removeHeadersService = new RemoveHeadersService(mockFilteringLog);
 
-    const context = {
-        requestUrl: 'https://example.org',
-        referrerUrl: 'https://example.org',
-        requestType: RequestType.Document,
-        contentType: ContentType.Document,
-        tabId: 0,
-        frameId: 0,
-        requestFrameId: 0,
-        timestamp: Date.now(),
-        thirdParty: false,
-        matchingResult: new MatchingResult([], null),
-        requestHeaders: [{
-            name: 'test_name',
-            value: 'test_value',
-        }],
-        responseHeaders: [{
-            name: 'test_name',
-            value: 'test_value',
-        }],
+    const getContext = (): RequestContext => {
+        return {
+            requestUrl: 'https://example.org',
+            referrerUrl: 'https://example.org',
+            requestType: RequestType.Document,
+            contentType: ContentType.Document,
+            tabId: 0,
+            frameId: 0,
+            requestFrameId: 0,
+            timestamp: Date.now(),
+            thirdParty: false,
+            matchingResult: new MatchingResult([], null),
+            requestHeaders: [{
+                name: 'test_name',
+                value: 'test_value',
+            }],
+            responseHeaders: [{
+                name: 'test_name',
+                value: 'test_value',
+            }],
+        } as RequestContext;
     };
 
+    let context = getContext();
+
     const runOnBeforeSendHeaders = (): boolean => {
-        return headersService.onBeforeSendHeaders(context as RequestContext);
+        return removeHeadersService.onBeforeSendHeaders(context);
     };
 
     const runOnHeadersReceived = (): boolean => {
-        return headersService.onHeadersReceived(context as RequestContext);
+        return removeHeadersService.onHeadersReceived(context);
     };
 
     beforeEach(() => {
+        context = getContext();
         mockFilteringLog.publishEvent.mockClear();
     });
 
     it('checks removing request headers', () => {
-        let headersModified = headersService.onBeforeSendHeaders(context as RequestContext);
+        let headersModified = removeHeadersService.onBeforeSendHeaders(context);
         expect(headersModified).toBeFalsy();
         expect(mockFilteringLog.publishEvent).not.toHaveBeenCalledWith(
             expect.objectContaining({ type: FilteringEventType.RemoveHeader }),
@@ -82,7 +87,7 @@ describe('Headers service', () => {
     });
 
     it('checks removing response headers', () => {
-        let headersModified = headersService.onHeadersReceived(context as RequestContext);
+        let headersModified = removeHeadersService.onHeadersReceived(context);
         expect(headersModified).toBeFalsy();
         expect(mockFilteringLog.publishEvent).not.toHaveBeenCalledWith(
             expect.objectContaining({ type: FilteringEventType.RemoveHeader }),
@@ -118,6 +123,28 @@ describe('Headers service', () => {
         headersModified = runOnHeadersReceived();
         expect(headersModified).toBeTruthy();
         expect(mockFilteringLog.publishEvent).toHaveBeenCalledWith(
+            expect.objectContaining({ type: FilteringEventType.RemoveHeader }),
+        );
+    });
+
+    it('correctly applies matching header modifier rules', () => {
+        context.matchingResult = new MatchingResult([
+            new NetworkRule('||example.org^$header=test_name:test_value,removeheader=test_name', 0),
+        ], null);
+        const headersModified = removeHeadersService.onHeadersReceived(context);
+        expect(headersModified).toBeTruthy();
+        expect(mockFilteringLog.publishEvent).toHaveBeenCalledWith(
+            expect.objectContaining({ type: FilteringEventType.RemoveHeader }),
+        );
+    });
+
+    it('does not apply non-matching header modifier rules', () => {
+        context.matchingResult = new MatchingResult([
+            new NetworkRule('||example.org^$header=test_name:NOT_test_value,removeheader=test_name', 0),
+        ], null);
+        const headersModified = removeHeadersService.onHeadersReceived(context);
+        expect(headersModified).toBeFalsy();
+        expect(mockFilteringLog.publishEvent).not.toHaveBeenCalledWith(
             expect.objectContaining({ type: FilteringEventType.RemoveHeader }),
         );
     });