+++ packages/tsurlfilter/CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Added
 - New `$permissions` modifier to set Permissions-Policy response header.
+- New `$header` modifier to match requests by response headers.
 - Support conversion to DNR for `$cookie` modifier.
 - `$url` modifier support for non-basic rules.